cert: move sss_cert_derb64_to_ldap_filter() out of libsss_cert.so
To avoid a dependency to libsss_certmap.so
sss_cert_derb64_to_ldap_filter() is removed from the internal
libsss_cert.so library.

Resolves: https://github.com/SSSD/sssd/issues/1030
